fix(#218): 9 second-wave CodeRabbit fixes — F16/BF16 conversion, legacy .pt bytes, TB hardening
Critical:
- PtReader legacy format: REDUCE was snapshotting RawStorage.Bytes
  while it was still null (legacy trailer hadn't been read yet) and
  baking that empty payload into PtTensorRef. Result: every legacy
  .pt tensor came back with zero bytes, even when the trailer fill
  succeeded. Fix: emit a DeferredLegacyTensor placeholder during
  pickle interp that captures the RawStorage REFERENCE; after the
  trailer fills the bytes, walk the result via MaterialiseDeferred
  and replace each placeholder with a concrete PtTensorRef. Modern
  zip format already loaded bytes eagerly in the persistent-id
  resolver, so it goes through the direct PtTensorRef path
  unchanged.

Major:
- AiDotNet.Tensors.Cli convert pt→safetensors: HalfStorage and
  BFloat16Storage fell into the default skip-and-warn branch.
  Almost every public HuggingFace checkpoint is F16 or BF16, so a
  conversion of e.g. Llama-2-7B would emit an empty safetensors
  file. Fix: route those through writer.AddRaw with the
  matching safetensors dtype tag (F16 / BF16) — lossless byte-for-
  byte transfer, no CLR Half/BFloat16 dependency. Adds
  ExtractContiguousBytes helper that handles both contiguous and
  strided source layouts.
- DistributedCheckpoint.PersistableTensor<T>.ByteCount: cast
  Tensor<T>.Length (int) to long BEFORE multiplying by
  MarshalSize(). The previous int×int overflowed for any tensor
  bigger than ~2 GB / sizeof(T), wrapping to a negative long that
  the sharded writer's bin-packing accepted as a positive shard
  size and skewed the layout.
- SafetensorsTensorEntry.Shape: expose as a clone-on-access
  property backed by a private array. Previously the getter
  returned the stored array directly, so external mutation could
  leave the entry inconsistent (ElementCount is cached at
  construction; mutating Shape would diverge from the cached
  count and skew downstream byte-length calculations).
- ShardedSafetensorsWriter.Add / AddRaw: mirror SafetensorsWriter's
  empty-name and reserved-name-"__metadata__" rejections. The
  per-shard SafetensorsWriter rejects them at Save time, but the
  buffered API used to accept them — so a malformed name would
  buffer OK and then Save() throw partway through after some
  shard files were already written. Fail fast at the buffered
  add-time instead.
- TensorBoardSummaryWriter.OpenLogDir: the
  events.out.tfevents.{unixSeconds}.{host} naming can collide for
  two writers opened in the same second on the same host, and
  FileMode.Create truncates the earlier run's log. Fix: use
  FileMode.CreateNew with a uniqueness-suffix loop
  (events.out.tfevents.<sec>.<host>.<n>) so a collision picks the
  next free filename instead of overwriting.

Minor:
- TensorBoardSummaryWriter.AddHistogram: handle max == min as a
  single-bucket delta-function histogram (one bucket whose upper
  limit is the value itself, count = N). The previous "fake
  width = 1" path invented a range that was never present and made
  TensorBoard render a spread-out histogram for what is actually
  a constant.
- TensorBoardSummaryWriter.AddHParam / LogHParams: explicit doc
  rewrite clarifying that these emit ordinary scalar summaries
  under hparams/* tags, NOT TensorBoard HParams plugin records.
  TensorBoard's HParams dashboard requires session_start_info /
  session_end_info / experiment-config metadata summaries; without
  those it doesn't recognise hparams/*-prefixed scalars as HParams
  data. Rather than mislead users into thinking the dashboard will
  pick these up, make the surface honest. Also add ThrowIfDisposed
  to LogHParams (the empty-dict short-circuit previously bypassed
  the disposed-state contract).

All 78 prior tests pass on net10.0.

+++ src/AiDotNet.Tensors/Serialization/Distributed/DistributedCheckpoint.cs
@@ -165,7 +165,12 @@ public sealed class PersistableTensor<T> : IPersistableTensor where T : struct
     }
 
     /// <inheritdoc />
-    public long ByteCount => _t.Length * MarshalSize();
+    // Widen Length to long BEFORE the multiply — Tensor<T>.Length is
+    // int and MarshalSize() returns int, so int × int overflows for
+    // any tensor of more than ~2 GB / sizeof(T). The previous form
+    // wrapped that overflow into a negative long, badly skewing the
+    // sharded writer's bin-packing.
+    public long ByteCount => (long)_t.Length * MarshalSize();
 
     /// <inheritdoc />
     public void AddTo(ShardedSafetensorsWriter writer, string name) => writer.Add(name, _t);
